WIP encryption
The rest of the encryption functions.

Still todo:

* Add tests for this.
* Add declarations to header files.
* Add error handling.
* General cleanup of the code, maybe add some comments.

The authkey is used to generate/verify the authenticator attribute
(by taking hmac of the full M1 and M2 WSC blocks.

It is also used to generate/verify the key_wrap_auth attribute that
comes at the end of the encrypted settings, by taking hmac of the
plaintext of encrypted settings (excluding key_wrap_auth).

keywrapkey is used to encrypt/decrypt the encrypted settings block
(including key_wrap_auth).